third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/109-expected.txt

This is a testharness.js-based test.
FAIL  scheduler: javascript URL in iframe, src set via DOM assert_array_equals: property 2, expected "inline script #3" but got "iframe onload"
Harness: the test ran to completion.
